# created a way to measure the performance of the server