Allows you to measure how many ticks per second the server has.

Example output:

> [5c6b2e5c][FPS]: 192fps AverageTickTime: 0.052768ms LongestTick: 0.367376ms
> [5c6b2e66][FPS]: 192fps AverageTickTime: 0.051188ms LongestTick: 0.267270ms
> [5c6b2e70][FPS]: 192fps AverageTickTime: 0.050705ms LongestTick: 0.267536ms
> [5c6b2e7a][FPS]: 192fps AverageTickTime: 0.052260ms LongestTick: 0.224877ms
> [5c6b2e84][FPS]: 192fps AverageTickTime: 0.051239ms LongestTick: 0.126669ms

Right now every 10 seconds 1 line be logged