@teoman002 2019-02-18
Is the place where you measure relevant?
For example does the code return the same fps if I put it into gameworld.CPP ontick() function?

Why is it defined in server.cpp?