@duralakun 2019-02-19
I think this is the main loop of the server, inside of it `GameServer()->OnTick();` will also be called.
But you could easily create another CMeasureTicks anywhere else in the code to measure that part.
You can do it like this:
```
CMeasureTicks MeasureTicks(10, "Tag"); // to create an objec, 10 second interval log
MeasureTicks.Begin(); // when the tick starts
MeasureTicks.End(); // when the tick ends
```