@duralakun 2019-02-19
After last commit, the output now looks like this for me:

> [5c6b8e1d][ServerTick]: 192fps, AverageTickTime: 5.192064ms, LongestTick: 5.486300ms
> [5c6b8e27][ServerTick]: 192fps, AverageTickTime: 5.188971ms, LongestTick: 5.518902ms
> [5c6b8e31][ServerTick]: 192fps, AverageTickTime: 5.201708ms, LongestTick: 5.292991ms

You can see the AverageTickTime is by 5ms longer.
It is because i moved the end of tick function down to include the network waiting, which takes 5ms.
This makes it less confusing, now 5.201708ms * 192fps = 1 second