@teoman002 2019-02-19
i7 6700k 4.7 Ghz
[5c6bb777][GameworldTick]: 50fps, AverageTickTime: 0.403353ms, LongestTick: 2.790882ms
[5c6bb781][GameworldTick]: 50fps, AverageTickTime: 0.415893ms, LongestTick: 1.619961ms
[5c6bb78b][GameworldTick]: 50fps, AverageTickTime: 0.508913ms, LongestTick: 4.168858ms
[5c6bb795][GameworldTick]: 50fps, AverageTickTime: 0.444099ms, LongestTick: 5.412899ms
[5c6bb79f][GameworldTick]: 50fps, AverageTickTime: 0.278836ms, LongestTick: 5.439547ms
[5c6bb7a9][GameworldTick]: 50fps, AverageTickTime: 0.616413ms, LongestTick: 4.481750ms
[5c6bb7b3][GameworldTick]: 50fps, AverageTickTime: 0.420286ms, LongestTick: 3.696510ms
[5c6bb7bd][GameworldTick]: 50fps, AverageTickTime: 0.064765ms, LongestTick: 0.590548ms