@duralakun 2019-02-20
This should save to merge, because this feature is disabled by default, only when you uncomment defines it will be activated. The only other thing this pr does is deleting obsolete code. I searched with grep for uses of it and couldn't find any.